+++ src/jtl/Connector/Shopware/Mapper/Product.php
@@ -183,6 +183,7 @@ class Product extends DataMapper
         $detailSW = null;
         //$result = new ProductModel();
         $result = $product;
+        $attrMappings = [];
 
         /*
         Logger::write(sprintf('>>> Product with id (%s, %s), masterProductId (%s, %s), manufacturerId (%s, %s)',
@@ -203,7 +204,7 @@ class Product extends DataMapper
 
                 $this->prepareChildAssociatedData($product, $productSW, $detailSW);
                 $this->prepareDetailAssociatedData($product, $productSW, $detailSW, true);
-                $this->prepareAttributeAssociatedData($product, $productSW, $detailSW, true);
+                $this->prepareAttributeAssociatedData($product, $productSW, $detailSW, $attrMappings, true);
                 $this->preparePriceAssociatedData($product, $productSW, $detailSW);
                 $this->prepareUnitAssociatedData($product, $detailSW);
                 $this->prepareMeasurementUnitAssociatedData($product, $detailSW);
@@ -231,7 +232,7 @@ class Product extends DataMapper
                 $this->prepareDetailAssociatedData($product, $productSW, $detailSW);
                 $this->prepareVariationAssociatedData($product, $productSW);
                 $this->prepareSpecificAssociatedData($product, $productSW, $detailSW);
-                $this->prepareAttributeAssociatedData($product, $productSW, $detailSW);
+                $this->prepareAttributeAssociatedData($product, $productSW, $detailSW, $attrMappings);
                 $this->preparePriceAssociatedData($product, $productSW, $detailSW);
                 $this->prepareUnitAssociatedData($product, $detailSW);
                 $this->prepareMeasurementUnitAssociatedData($product, $detailSW);
@@ -256,7 +257,7 @@ class Product extends DataMapper
                 $this->prepareSetVariationRelations($product, $productSW);
                 $this->saveVariationTranslationData($product, $productSW);
                 $this->deleteTranslationData($productSW);
-                $this->saveTranslationData($product, $productSW);
+                $this->saveTranslationData($product, $productSW, $attrMappings);
             }
         } catch (\Exception $e) {
             Logger::write(sprintf('Exception from Product (%s, %s)', $product->getId()->getEndpoint(), $product->getId()->getHost()), Logger::ERROR, 'database');
@@ -606,7 +607,7 @@ class Product extends DataMapper
         }
     }
 
-    protected function prepareAttributeAssociatedData(ProductModel $product, ArticleSW &$productSW, DetailSW &$detailSW, $isChild = false)
+    protected function prepareAttributeAssociatedData(ProductModel $product, ArticleSW &$productSW, DetailSW &$detailSW, array &$attrMappings, $isChild = false)
     {
         // Attribute
         $attributeSW = $detailSW->getAttribute();
@@ -625,6 +626,7 @@ class Product extends DataMapper
         }
 
         $i = 3;
+        $attrMappings = [];
         foreach ($product->getAttributes() as $attribute) {
             if (!$attribute->getIsCustomProperty()) {
                 $i++;
@@ -660,7 +662,11 @@ class Product extends DataMapper
 
                                     if ($number != $i && method_exists($attributeSW, $setter)) {
                                         $attributeSW->{$setter}($oldValue);
+                                        $hostId = $attrMappings[$number];
+                                        $attrMappings[$i] = $hostId;
                                     }
+
+                                    $attrMappings[$number] = $attribute->getId()->getHost();
 
                                     continue;
                                 }
@@ -669,6 +675,7 @@ class Product extends DataMapper
 
                         if (method_exists($attributeSW, $setter)) {
                             $attributeSW->{$setter}($attributeI18n->getValue());
+                            $attrMappings[$i] = $attribute->getId()->getHost();
                         }
                     }
                 }
@@ -871,10 +878,28 @@ class Product extends DataMapper
         $productSW->setPropertyGroup($group);
     }
 
-    protected function saveTranslationData(ProductModel $product, ArticleSW $productSW)
+    protected function saveTranslationData(ProductModel $product, ArticleSW $productSW, array $attrMappings)
     {
-        // ProductI18n
         $shopMapper = Mmc::getMapper('Shop');
+
+        // AttributeI18n
+        $attrI18ns = [];
+        foreach ($product->getAttributes() as $attr) {
+            foreach ($attr->getI18ns() as $attrI18n) {
+                if ($attrI18n->getLanguageISO() !== LanguageUtil::map(Shopware()->Shop()->getLocale()->getLocale())) {
+                    if (!isset($attrI18ns[$attrI18n->getLanguageISO()])) {
+                        $attrI18ns[$attrI18n->getLanguageISO()] = [];
+                    }
+
+                    if (($index = array_search($attr->getId()->getHost(), $attrMappings)) !== false) {
+                        $i = "attr{$index}";
+                        $attrI18ns[$attrI18n->getLanguageISO()][$i] = $attrI18n->getValue();
+                    }
+                }
+            }
+        }
+
+        // ProductI18n
         $translationUtil = new TranslationUtil();
         $cache = array();
         foreach ($product->getI18ns() as $i18n) {
@@ -904,6 +929,10 @@ class Product extends DataMapper
                             'attr2' => '',
                             'attr3' => ''
                         );
+
+                        if (isset($attrI18ns[$i18n->getLanguageISO()])) {
+                            $cache[$shop->getId()] = array_merge($cache[$shop->getId()], $attrI18ns[$i18n->getLanguageISO()]);
+                        }
 
                         $translationUtil->write(
                             $shop->getId(),